move exception under handler
