+++ src/main/java/ca/uwaterloo/drinkmasterapi/feature/order/service/OrderServiceImpl.java
@@ -1,6 +1,6 @@
 package ca.uwaterloo.drinkmasterapi.feature.order.service;
 
-import ca.uwaterloo.drinkmasterapi.common.ResourceNotFoundException;
+import ca.uwaterloo.drinkmasterapi.handler.exception.ResourceNotFoundException;
 import ca.uwaterloo.drinkmasterapi.dao.Drink;
 import ca.uwaterloo.drinkmasterapi.repository.DrinkRepository;
 import ca.uwaterloo.drinkmasterapi.repository.MachineRepository;
