+++ src/main/java/ca/uwaterloo/drinkmasterapi/feature/user/service/UserLoginServiceImpl.java
@@ -1,6 +1,6 @@
 package ca.uwaterloo.drinkmasterapi.feature.user.service;
 
-import ca.uwaterloo.drinkmasterapi.common.InvalidCredentialsException;
+import ca.uwaterloo.drinkmasterapi.handler.exception.InvalidCredentialsException;
 import ca.uwaterloo.drinkmasterapi.feature.user.dto.LoginRequestDTO;
 import ca.uwaterloo.drinkmasterapi.feature.user.dto.UserResponseDTO;
 import ca.uwaterloo.drinkmasterapi.feature.user.dto.SignupRequestDTO;
