+++ src/main/java/ca/uwaterloo/drinkmasterapi/handler/exception/InvalidCredentialsException.java
@@ -1,4 +1,4 @@
-package ca.uwaterloo.drinkmasterapi.common;
+package ca.uwaterloo.drinkmasterapi.handler.exception;
 
 public class InvalidCredentialsException extends RuntimeException {
     public InvalidCredentialsException(String message) {
